fix: corrigir 403 causado por type mismatch bigint/int no PostgreSQL
O PDO_PGSQL retorna colunas bigint como string PHP quando
ATTR_EMULATE_PREPARES está ativo. O comparison strict (!==) entre
$model->tenant_id (string "4") e app('tenant')->id (int 4) resultava
em true, disparando abort(403) incorretamente.

Correções:
- Adicionar tenant_id => integer no $casts dos models afetados
- Adicionar cast (int) explícito em todos os abort_if de tenant
- Remover logging de debug temporário

+++ app/Http/Controllers/Tenant/ClienteController.php
@@ -31,7 +31,7 @@ class ClienteController extends Controller
 
     public function show(Cliente $cliente): Response
     {
-        abort_if($cliente->tenant_id !== app('tenant')->id, 403);
+        abort_if((int)$cliente->tenant_id !== (int)app('tenant')->id, 403);
 
         return Inertia::render('Tenant/Clientes/Show', [
             'cliente'      => $cliente,
